Add an AudioConfig
The AudioConfig allows to edit real/virtual voices, even at runtime (tested with LethalConfig), which is useful to tweak it in game or reduce these numbers when the audio becomes overwhelming or performances can't keep up.
The config take a list of acceptable values which comes from Unity documentation and takes the closest value if the one entered isn't valid (instead of the first value by default).
I also updated the project description and version.